rm conj op
