+++ oneflow/ir/include/OneFlow/OneFlowUserOps.td
@@ -11807,33 +11807,6 @@ def OneFlow_ImagGradOp : OneFlow_BaseOp<"imag_grad", [NoGrad, DeclareOpInterface
   let has_data_type_infer_fn = 1;
 }
 
-def OneFlow_ConjOp : OneFlow_BaseOp<"conj", [NoGrad, DeclareOpInterfaceMethods<UserOpCompatibleInterface>]> {
-  let input = (ins
-    OneFlow_Tensor:$x
-  );
-  let output = (outs
-    OneFlow_Tensor:$out
-  );
-  let has_logical_tensor_desc_infer_fn = 1;
-  let has_physical_tensor_desc_infer_fn = 1;
-  let has_get_sbp_fn = 1;
-  let has_data_type_infer_fn = 1;
-}
-
-def OneFlow_ConjGradOp : OneFlow_BaseOp<"conj_grad", [NoGrad, DeclareOpInterfaceMethods<UserOpCompatibleInterface>]> {
-  let input = (ins
-    OneFlow_Tensor:$dout
-    OneFlow_Tensor:$x,
-  );
-  let output = (outs
-    OneFlow_Tensor:$dx
-  );
-  let has_logical_tensor_desc_infer_fn = 1;
-  let has_physical_tensor_desc_infer_fn = 1;
-  let has_get_sbp_fn = 1;
-  let has_data_type_infer_fn = 1;
-}
-
 def OneFlow_ConjPhysicalOp : OneFlow_BaseOp<"conj_physical", [NoGrad, DeclareOpInterfaceMethods<UserOpCompatibleInterface>]> {
   let input = (ins
     OneFlow_Tensor:$x
