+++ oneflow/user/kernels/complex_kernels_util.cu
@@ -15,7 +15,7 @@ limitations under the License.
 */
 #ifdef WITH_CUDA
 #include "oneflow/core/framework/framework.h"
-#include "oneflow/user/kernels/real_kernel_util.h"
+#include "oneflow/user/kernels/complex_kernels_util.h"
 #include <cufft.h>
 
 namespace oneflow {
@@ -31,6 +31,26 @@ struct RealFunctor<DeviceType::kCUDA, dtype_x, dtype_out> final {
 
 INSTANTIATE_REAL_FUNCTOR(DeviceType::kCUDA, cufftComplex, float)
 INSTANTIATE_REAL_FUNCTOR(DeviceType::kCUDA, cufftComplexDouble, double)
+
+template<typename dtype_x, typename dtype_out>
+struct ImagFunctor<DeviceType::kCUDA, dtype_x, dtype_out> final {
+  void operator()(ep::Stream* stream, const dtype_x* x, const dtype_out* out) {
+    // TODO(lml): finish this function.
+  }
+};
+
+INSTANTIATE_IMAG_FUNCTOR(DeviceType::kCUDA, cufftComplex, float)
+INSTANTIATE_IMAG_FUNCTOR(DeviceType::kCUDA, cufftComplexDouble, double)
+
+template<typename dtype>
+struct ConjPhysicalFunctor<DeviceType::kCUDA, dtype> final {
+  void operator()(ep::Stream* stream, const dtype* x, const dtype* out) {
+    // TODO(lml): finish this function.
+  }
+};
+
+INSTANTIATE_CONJ_PHYSICAL_FUNCTOR(DeviceType::kCUDA, cufftComplex)
+INSTANTIATE_CONJ_PHYSICAL_FUNCTOR(DeviceType::kCUDA, cufftComplexDouble)
 
 }  // namespace user_op
 }  // namespace oneflow
