--- oneflow/user/ops/conj_op.cpp
@@ -1,48 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#include "oneflow/core/framework/framework.h"
-#include "oneflow/core/framework/op_generated.h"
-
-namespace oneflow {
-// TODO(lml): add infer is_conj flag
-
-/*static*/ Maybe<void> ConjOp::GetSbp(user_op::SbpContext* ctx) {
-  return user_op::GetSbpFnUtil::SplitForEachAxis(ctx);
-}
-/*static*/ Maybe<void> ConjOp::InferLogicalTensorDesc(user_op::InferContext* ctx) {
-  return user_op::TensorDescInferFnUtil::Unchanged(ctx);
-}
-/*static*/ Maybe<void> ConjOp::InferPhysicalTensorDesc(user_op::InferContext* ctx) {
-  return InferLogicalTensorDesc(ctx);
-}
-/*static*/ Maybe<void> ConjOp::InferDataType(user_op::InferContext* ctx) {
-  return user_op::TensorDescInferFnUtil::UnchangedDataType(ctx);
-}
-
-/*static*/ Maybe<void> ConjGradOp::GetSbp(user_op::SbpContext* ctx) {
-  return user_op::GetSbpFnUtil::SplitForEachAxis(ctx);
-}
-/*static*/ Maybe<void> ConjGradOp::InferLogicalTensorDesc(user_op::InferContext* ctx) {
-  return user_op::TensorDescInferFnUtil::Unchanged(ctx);
-}
-/*static*/ Maybe<void> ConjGradOp::InferPhysicalTensorDesc(user_op::InferContext* ctx) {
-  return InferLogicalTensorDesc(ctx);
-}
-/*static*/ Maybe<void> ConjGradOp::InferDataType(user_op::InferContext* ctx) {
-  return user_op::TensorDescInferFnUtil::UnchangedDataType(ctx);
-}
-
-}  // namespace oneflow
